+++ scifit/solvers/specials.py
@@ -1,0 +1,5 @@
+import numpy as np
+
+from scifit.interfaces.generic import FitSolverInterface
+
+